requirements.txt
fastapi==0.89.1
uvicorn==0.17.3
requests==2.28.2
pydantic==1.9.0
python-multipart==0.0.6
python-jose==3.3.0